Fix filename checks wrt basenames in read_fts()
Fix bug in read_fts() with checking basename versus full path. In particular
if the number of directories is just one then it has to be a basename check.
Thus it's (fts->base || count_dirs(name) == 1) && .... If it's not base or the
count_dirs(name) != 1 then it tries the other way (with fts->base == false).
This solves a problem where files to be ignored were not ignored in the case of
base being false (when we needed that to be the case).