Save and restore errno in pipe_open()/shell_cmd()
Save and restore errno prior to returning from pipe_open() and
shell_cmd().  In some systems popen(3) are not reliable with errno
(like macOS) but in linux it CAN in some cases set errno. For
system(3) it does no seem to set errno at all but we still do this
process especially as it also uses a function that can have errno (and
we return from the function in that case with an error).