Additional sanity checks added to FTS code
The read_fts() function now does not just skip on NULL or empty path name; it
will abort just like check_fts_info() does (though that would never have been
reached as we skipped that item and moved to the next). If there is a NULL
pointer or an empty string something is wrong and it's better to abort early.

Also in check_fts_info() when the fts_info == FTS_NS the ent->fts_errno is
the errno so we set errno = ent->fts_errno prior to calling errp() (instead
of err()).

Updated JPARSE_UTILS_VERSION to "2.0.3 2025-03-07".