dir_name()/base_name() updates and new function
The function base_name() now will, on a NULL or empty string, return a
copy of "." just like basename(3) does.

The function dir_name() now will, on a NULL or empty string or a string
that has no / (unless level < 0) , return a copy of "." just like
dirname(3) (though that function does not have the level functionality).

The function resolve_path() will go through the process's $PATH and
search for an executable file with the name passed to the function,
returning either a copy of the path or NULL. If the command starts with
/ or ./ then the $PATH is not searched and instead that exact string is
duplicated and returned.