Add setlocale(LC_ALL, ""); to all main() functions
After a discussion and review now early on in every main() we do a
setlocale(LC_ALL, ""); so that the default locale (in the system) based
on LANG will be used.

NOTE: this is not done in the library for very good reasons.

Updated JSEMTBLGEN_VERSION to "2.0.1 2025-03-15".
Updated JSTRDECODE_VERSION to "2.2.2 2025-03-15".
Updated JSTRENCODE_VERSION to "2.2.2 2025-03-15".
Updated JNUM_CHK_VERSION to "2.0.1 2025-03-15".
Updated JNUM_GEN_VERSION to "2.0.1 2025-03-15".
Updated PR_JPARSE_TEST_VERSION to "2.0.1 2025-03-15".
Updated UTIL_TEST_VERSION to "2.0.3 2025-03-15".
Updated VERGE_VERSION to "2.0.2 2025-03-15".
Updated JPARSE_TOOL_VERSION to "2.0.1 2025-03-15".