Add fnmatch(3) support to ignore list for read_fts
Added fnmatch_flags to struct fts for ignored list of files. This is by
request for mkiocccentry. It is possibly feasible to have it for non-ignored
paths but that is not for now if it will be done (or even desired).

Removed a test in util_test because it is faulty in logic due to depth. A test
for fnmatch(3) was added.

Updated JPARSE_UTILS_VERSION to  "2.0.10 2025-06-19".
Updated UTIL_TEST_VERSION to "2.0.5 2025-06-19".