Fix CHANGES.md for previous commit
The following was meant to be added:

It is entirely unclear if every case of 'true'/'false' is correct in
every case due to the fact that this update is mostly for the fts(3)
routines and there is an option to ignore and also an option to match,
both of which can accept globs.  If a problem arises the call to these
functions can be revised but this does resolve a problem observed in the
need for the fnmatch(3) change.