Undo accidental commit
In the previous commit only the man page for jparse_bug_report.sh and
jparse_bug_report.sh itself were meant to be committed. The other code
was a WIP that was not even working (it was haphazardly written, is
incorrect and has to be looked at when there is more time and after some
discussion of what to do with raw bytes).