increase debug message levels in jparse/util.c and jparse/verge.c
Increased the debug level of most debug messages for functions in `jparse/verge.c`.
In some cases a debug message was turned into a warning when an non-fatal
error condition was detected.

Increased the debug level of non-UTIL_TEST code debug messages for
functions in `jparse/util.c`.

Updated `JPARSE_UTILS_VERSION` to `"2.0.15 2025-09-01"`.