jparse(1) now allows multiple args
Improved jparse(1) to allow for more than one arg of the specific type
(string if -s is used, file otherwise). If any of the args is invalid
JSON it will exit 1, else 0.

Updated error location files as the jparse(1) error now specifies the
filename of each invalid JSON file. This is important for the test
script.

Updated JPARSE_TOOL_VERSION to "2.0.2 2025-09-09".