Removed unused macros from util.h
Removed unused macros from util.h, some of which are leftovers from
mkiocccentry (at the time of moving code out of here that never belonged
here but was here due to the history of the repo the macros were
missed).

Updated JPARSE_UTILS_VERSION to "2.1.3 2025-09-20" (yes this jumped
up but that's because the last two code updates to util.[ch] did not
have a version update).